+++ gommander_test.go
@@ -14,20 +14,22 @@ func TestCommandMetadata(t *testing.T) {
 	assertEq(t, cmd.GetHelp(), "Basic command", "Cmd help field set incorrectly")
 	assertEq(t, cmd.GetAuthor(), "vndaba", "Cmd author set wrongly")
 	assertEq(t, cmd.GetVersion(), "0.1.0", "Cmd Version set wrongly")
-	assertArrEq(t, cmd.GetAliases(), []string{"b"}, "Cmd aliases set wrongly")
-	assertStructEq[*Flag](t, cmd.GetFlags()[0], helpFlag(), "Help flag not added automatically")
+	assertDeepEq(t, cmd.GetAliases(), []string{"b"}, "Cmd aliases set wrongly")
+	assertDeepEq(t, cmd.GetFlags()[0], helpFlag(), "Help flag not added automatically")
 }
 
 func TestCommandSettings(t *testing.T) {
+	clearCache()
 	app := App()
-	app.SubCommand("dummy")
+	app.SubCommand("dummy").Argument("<uint:count>", "a count arg")
 
 	// assertStructArrEq[*Flag](t, app.GetFlags(), []*Flag{helpFlag(), versionFlag()}, "Help and version flags not set correctly")
 	assertEq(t, len(app.GetFlags()), 2, "Help and version flags not set correctly")
-	assertStructEq[*Flag](t, app.GetFlags()[1], versionFlag(), "Version flag not set correctly")
+	assertDeepEq(t, app.GetFlags()[1], versionFlag(), "Version flag not set correctly")
 
 	app.Set(DisableVersionFlag, true)
 	app.Set(IncludeHelpSubcommand, true)
+	app.Set(AllowNegativeNumbers, true)
 	// TODO: Complete the other settings
 	app._init()