+++ help.go
@@ -30,19 +30,19 @@ func (HelpWriter) Write(c *Command) {
 	fmter.Add(Keyword, fmt.Sprintf("    %v", c._getUsageStr()))
 	if !hasCustomUsage {
 		if hasFlags {
-			fmter.Add(Other, " [FLAGS]")
+			fmter.Add(Other, " "+app.flagsHelpValue)
 		}
 
 		if hasOptions {
-			fmter.Add(Other, " [OPTIONS]")
+			fmter.Add(Other, " "+app.optionsHelpValue)
 		}
 
 		if hasArgs {
-			fmter.Add(Other, " <ARGS>")
+			fmter.Add(Other, " "+app.argsHelpValue)
 		}
 
 		if hasSubcmds {
-			fmter.Add(Other, " <SUBCOMMAND>")
+			fmter.Add(Other, " "+app.subCmdsHelpValue)
 		}
 	}
 	fmter.close()
@@ -53,22 +53,22 @@ func (HelpWriter) Write(c *Command) {
 	}
 
 	if hasArgs {
-		fmter.section("ARGS")
+		fmter.section(app.argsHelpHeading)
 		fmter.format(standardize(c.arguments))
 	}
 
 	if hasFlags {
-		fmter.section("FLAGS")
+		fmter.section(app.flagsHelpHeading)
 		fmter.format(standardize(c.flags))
 	}
 
 	if hasOptions {
-		fmter.section("OPTIONS")
+		fmter.section(app.optionsHelpHeading)
 		fmter.format(standardize(c.options))
 	}
 
 	if hasSubcmds && !hasSubcmdGroups {
-		fmter.section("SUBCOMMANDS")
+		fmter.section(app.subCmdsHelpHeading)
 		fmter.format(standardize(c.subCommands))
 	}