+++ CHANGELOG.md
@@ -7,6 +7,29 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Updated the Readme documentation to include the newly added functionality for arg types and methods.
+- Added new predefined argument types: `file` for filenames and directories and `uint` for unsigned integers
+- Enabled validation of argument values using regular expressions via the `Arg.ValidatorRegex()` method
+- Integrated the `AllowNegativeNumbers` setting into the program
+- Enabled customization of the help headings and help values to be used for sub-cmds, arguments, flags and options.
+- Implemented a simple caching policy to improve performance
+
+### Changed
+
+- Improved general package performance and structs creation latency
+- Updated existing and added more benchmarks
+- Modified some of the error messages to be more concise
+- Renamed struct fields for Options, Arguments and Flags to make them public and enable creation using constructors
+- Modified the `FormatGenerator` interface to take in a pointer to the app for settings integration
+- Moved the suggest subcommands functionality from utils to a method of the `Command` struct
+
+### Removed
+
+- Removed unused assertion utility functions such as the assertArrEq and assertStructEq.
+- Removed the `.compare()` method for the StructComp interface implementations. Now comparing structs using the reflect.DeepEqual method.
+
 ## [0.2.0] - 2022-06-30
 
 ### Added